pyproject.toml
[project]
name = "dexp"
version = "0.1.0"
description = "Logistics Data Explorer"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ['streamlit>=1.0.0',
                'pandas>=1.0.0',
                'numpy>=1.18.0',
                'sqlalchemy>=1.4.0',
                'fastapi[standard]>=0.104.0',
                'uvicorn>=0.24.0',
                'crewai',
                'python-dotenv',
                'requests']

[project.scripts]
run_ui = "ux.src.__app__:run"
test_db = "database.src.__db__:test"
test_db_api = "database.src.__db_api__:test"
test_chat = "chat.src.__chat__:test"
test_chat_api = "chat.src.__chat_api__:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ux", "database", "chat"]